+++ core/dbt/task/base.py
@@ -26,10 +26,6 @@ from dbt.contracts.results import (
 )
 from dbt.common.events.contextvars import get_node_info
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
-    LogDebugStackTrace,
-    LogSkipBecauseError,
-)
 from dbt.events.types import (
     SkippingDetails,
     NodeCompiling,
@@ -38,6 +34,8 @@ from dbt.events.types import (
     InternalErrorOnRun,
     GenericExceptionOnRun,
     NodeConnectionReleaseError,
+    LogDebugStackTrace,
+    LogSkipBecauseError,
 )
 from dbt.common.exceptions import (
     DbtRuntimeError,
