+++ core/dbt/task/clean.py
@@ -3,7 +3,7 @@ from shutil import rmtree
 
 from dbt import deprecations
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
+from dbt.events.types import (
     CheckCleanPath,
     ConfirmCleanPath,
     FinishedCleanPaths,
