+++ core/dbt/task/list.py
@@ -10,8 +10,7 @@ from dbt.common.events.functions import (
     fire_event,
     warn_or_error,
 )
-from dbt.common.events.types import ListCmdOut
-from dbt.events.types import NoNodesSelected
+from dbt.events.types import NoNodesSelected, ListCmdOut
 from dbt.common.exceptions import DbtRuntimeError, DbtInternalError
 from dbt.common.events.contextvars import task_contextvars
