+++ core/dbt/task/printer.py
@@ -4,8 +4,8 @@ from dbt.logger import (
     TextOnly,
 )
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import (
-    Formatting,
+from dbt.common.events.types import Formatting
+from dbt.events.types import (
     RunResultWarning,
     RunResultWarningMessage,
     RunResultFailure,
