+++ core/dbt/task/run_operation.py
@@ -11,8 +11,8 @@ from dbt.contracts.files import FileHash
 from dbt.contracts.graph.nodes import HookNode
 from dbt.contracts.results import RunResultsArtifact, RunResult, RunStatus, TimingInfo
 from dbt.common.events.functions import fire_event
-from dbt.common.events.types import LogDebugStackTrace
 from dbt.events.types import (
+    LogDebugStackTrace,
     RunningOperationCaughtError,
     RunningOperationUncaughtError,
 )
