+++ tests/unit/test_events.py
@@ -24,7 +24,7 @@ from dbt.events.base_types import (
 from dbt.common.events.event_manager import TestEventManager, EventManager
 from dbt.common.events.functions import msg_to_dict, msg_to_json
 from dbt.common.events.helpers import get_json_string_utcnow
-from dbt.common.events.types import RunResultError
+from dbt.events.types import RunResultError
 from dbt.flags import set_from_args
 from dbt.task.printer import print_run_result_error
 
@@ -406,46 +406,46 @@ sample_values = [
     core_types.NodeConnectionReleaseError(node_name="", exc=""),
     core_types.FoundStats(stat_line=""),
     # Z - misc ======================
-    types.MainKeyboardInterrupt(),
-    types.MainEncounteredError(exc=""),
-    types.MainStackTrace(stack_trace=""),
+    core_types.MainKeyboardInterrupt(),
+    core_types.MainEncounteredError(exc=""),
+    core_types.MainStackTrace(stack_trace=""),
     types.SystemCouldNotWrite(path="", reason="", exc=""),
     types.SystemExecutingCmd(cmd=[""]),
     types.SystemStdOut(bmsg=str(b"")),
     types.SystemStdErr(bmsg=str(b"")),
     types.SystemReportReturnCode(returncode=0),
-    types.TimingInfoCollected(),
-    types.LogDebugStackTrace(),
-    types.CheckCleanPath(path=""),
-    types.ConfirmCleanPath(path=""),
-    types.ProtectedCleanPath(path=""),
-    types.FinishedCleanPaths(),
-    types.OpenCommand(open_cmd="", profiles_dir=""),
-    types.RunResultWarning(resource_type="", node_name="", path=""),
-    types.RunResultFailure(resource_type="", node_name="", path=""),
-    types.StatsLine(stats={"error": 0, "skip": 0, "pass": 0, "warn": 0, "total": 0}),
-    types.RunResultError(msg=""),
-    types.RunResultErrorNoMessage(status=""),
-    types.SQLCompiledPath(path=""),
-    types.CheckNodeTestFailure(relation_name=""),
-    types.EndOfRunSummary(num_errors=0, num_warnings=0, keyboard_interrupt=False),
-    types.LogSkipBecauseError(schema="", relation="", index=0, total=0),
-    types.EnsureGitInstalled(),
-    types.DepsCreatingLocalSymlink(),
-    types.DepsSymlinkNotAvailable(),
-    types.DisableTracking(),
-    types.SendingEvent(kwargs=""),
-    types.SendEventFailure(),
-    types.FlushEvents(),
-    types.FlushEventsFailure(),
+    core_types.TimingInfoCollected(),
+    core_types.LogDebugStackTrace(),
+    core_types.CheckCleanPath(path=""),
+    core_types.ConfirmCleanPath(path=""),
+    core_types.ProtectedCleanPath(path=""),
+    core_types.FinishedCleanPaths(),
+    core_types.OpenCommand(open_cmd="", profiles_dir=""),
+    core_types.RunResultWarning(resource_type="", node_name="", path=""),
+    core_types.RunResultFailure(resource_type="", node_name="", path=""),
+    core_types.StatsLine(stats={"error": 0, "skip": 0, "pass": 0, "warn": 0, "total": 0}),
+    core_types.RunResultError(msg=""),
+    core_types.RunResultErrorNoMessage(status=""),
+    core_types.SQLCompiledPath(path=""),
+    core_types.CheckNodeTestFailure(relation_name=""),
+    core_types.EndOfRunSummary(num_errors=0, num_warnings=0, keyboard_interrupt=False),
+    core_types.LogSkipBecauseError(schema="", relation="", index=0, total=0),
+    core_types.EnsureGitInstalled(),
+    core_types.DepsCreatingLocalSymlink(),
+    core_types.DepsSymlinkNotAvailable(),
+    core_types.DisableTracking(),
+    core_types.SendingEvent(kwargs=""),
+    core_types.SendEventFailure(),
+    core_types.FlushEvents(),
+    core_types.FlushEventsFailure(),
     types.Formatting(),
-    types.TrackingInitializeFailure(),
-    types.RunResultWarningMessage(),
-    types.DebugCmdOut(),
-    types.DebugCmdResult(),
-    types.ListCmdOut(),
+    core_types.TrackingInitializeFailure(),
+    core_types.RunResultWarningMessage(),
+    core_types.DebugCmdOut(),
+    core_types.DebugCmdResult(),
+    core_types.ListCmdOut(),
     types.Note(msg="This is a note."),
-    types.ResourceReport(),
+    core_types.ResourceReport(),
 ]
